+++ Changelog.md
@@ -38,3 +38,4 @@
 - When deleting a plugin that other plugins depend on, it will prompt you if you want to delete those as well.
 - Removed "showErrors" config option. A warning/error message will always display. However, you can disable the java errors with "showErrorTrace".
 - If a plugin version has multiple files it will prompt you for which ones to install.
+- Added "/plugin reload", which makes CogWorks rescan the "./plugins" folder for any changes.